+++ capitole-exam-business/src/main/java/com/edh/capitole/business/usecases/FindPriceRateAtDayUseCase.java
@@ -7,5 +7,5 @@ import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 public interface FindPriceRateAtDayUseCase {
-    Mono<PriceRateBo> execute(BigInteger productId, BigInteger brandId, LocalDateTime date);
+    Mono<PriceRateBo> execute(Long productId, Long brandId, LocalDateTime date);
 }
